backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary
python-decouple==3.8
django-cors-headers==4.4.0
openpyxl==3.1.5
django-import-export